improved support for partial website scraping
