+++ README.md
@@ -7,16 +7,30 @@ extract appearances as a JSON object.
 ## run
 This is a [`scrapy`](https://scrapy.org/) project, so it needs to be run with the
 `scrapy` command line util.
+### auto
+The `auto` spider recurses the Transfermarkt website hierarchy automatically for all reachable players. It can be invoked with the command
 ```console
 scrapy crawl auto
 ```
+> :warning: The `auto` spider scrapes the whole website hierarchy and therefore it can take quite some time to run. Check the `partial` spider for scoped website scrapping.
+### partial
+The `partial` spider uses a `SITE_MAP` setting to define the scrapping scope. The `SITE_MAP` 
+is a JSON representation of the site hierarchy that can be generated a single time by running
+```console
+scrapy crawl mapper > site_map.json
+```
+By using this site map generated with the `mapper` crawler above, a `SITE_MAP` setting can be populated now in the [settings.py](tfmkt/settings.py) to do partial scraping with
+```console
+scrapy crawl partial
+```
+In [this example](tfmkt/site_map.py), a `SITE_MAP` is provided that can be used to scrape player statistics from Premier League's Aston Vila. Additional Premier League clubs can be commented out from the site map in order to have those scraped as well.
 
 ## config
-The website hierarchy recursed can be trimmed by using the configuration `SITE_TREE_FILTER`.
+The website hierarchy recursed can be trimmed by using the configuration `SITE_MAP`.
 ```python
 # if passed, this setting will be used to limit the scope of the scraping
-# by filtering out paths from the site hierachy not defined here
-SITE_TREE_FILTER = {
+# by filtering out paths from the site hierachy that are not defined here
+SITE_MAP = {
     # confederation
     '/wettbewerbe/europa': {
         # competition
