+++ tfmkt/site_map.py
@@ -1,0 +1,638 @@
+# generate a site map by running
+# >>> scrapy crawl map
+
+site_map = {
+    "/wettbewerbe/europa": {
+        "/premier-league/startseite/wettbewerb/GB1": {
+            # "/aston-villa/startseite/verein/405/saison_id/2020": [
+            #     "henri-lansbury/leistungsdaten/spieler/44794/plus/1?saison=2018",
+            #     "matty-cash/leistungsdaten/spieler/425334/plus/1?saison=2018",
+            #     "ezri-konsa/leistungsdaten/spieler/413403/plus/1?saison=2018",
+            #     "anwar-el-ghazi/leistungsdaten/spieler/183720/plus/1?saison=2018",
+            #     "lovre-kalinic/leistungsdaten/spieler/46405/plus/1?saison=2018",
+            #     "conor-hourihane/leistungsdaten/spieler/61653/plus/1?saison=2018",
+            #     "bertrand-traore/leistungsdaten/spieler/131996/plus/1?saison=2018",
+            #     "frederic-guilbert/leistungsdaten/spieler/354352/plus/1?saison=2018",
+            #     "john-mcginn/leistungsdaten/spieler/193116/plus/1?saison=2018",
+            #     "marvelous-nakamba/leistungsdaten/spieler/324882/plus/1?saison=2018",
+            #     "douglas-luiz/leistungsdaten/spieler/447661/plus/1?saison=2018",
+            #     "jed-steer/leistungsdaten/spieler/110867/plus/1?saison=2018",
+            #     "bjorn-engels/leistungsdaten/spieler/193469/plus/1?saison=2018",
+            #     "jacob-ramsey/leistungsdaten/spieler/503749/plus/1?saison=2018",
+            #     "emiliano-martinez/leistungsdaten/spieler/111873/plus/1?saison=2018",
+            #     "tyrone-mings/leistungsdaten/spieler/253677/plus/1?saison=2018",
+            #     "trezeguet/leistungsdaten/spieler/234189/plus/1?saison=2018",
+            #     "ollie-watkins/leistungsdaten/spieler/324358/plus/1?saison=2018",
+            #     "keinan-davis/leistungsdaten/spieler/412660/plus/1?saison=2018",
+            #     "tom-heaton/leistungsdaten/spieler/34130/plus/1?saison=2018",
+            #     "jack-grealish/leistungsdaten/spieler/203460/plus/1?saison=2018",
+            #     "wesley/leistungsdaten/spieler/381362/plus/1?saison=2018",
+            #     "ahmed-elmohamady/leistungsdaten/spieler/66333/plus/1?saison=2018",
+            #     "kortney-hause/leistungsdaten/spieler/233124/plus/1?saison=2018",
+            #     "matt-targett/leistungsdaten/spieler/250478/plus/1?saison=2018",
+            #     "ross-barkley/leistungsdaten/spieler/131978/plus/1?saison=2018",
+            #     "neil-taylor/leistungsdaten/spieler/67416/plus/1?saison=2018"
+            # ],
+            # "/brighton-amp-hove-albion/startseite/verein/1237/saison_id/2020": [
+            #     "dan-burn/leistungsdaten/spieler/134270/plus/1?saison=2018",
+            #     "alireza-jahanbakhsh/leistungsdaten/spieler/213268/plus/1?saison=2018",
+            #     "pascal-gross/leistungsdaten/spieler/82873/plus/1?saison=2018",
+            #     "mathew-ryan/leistungsdaten/spieler/128969/plus/1?saison=2018",
+            #     "adam-lallana/leistungsdaten/spieler/43530/plus/1?saison=2018",
+            #     "neal-maupay/leistungsdaten/spieler/217115/plus/1?saison=2018",
+            #     "yves-bissouma/leistungsdaten/spieler/410425/plus/1?saison=2018",
+            #     "steven-alzate/leistungsdaten/spieler/476237/plus/1?saison=2018",
+            #     "adam-webster/leistungsdaten/spieler/212847/plus/1?saison=2018",
+            #     "lewis-dunk/leistungsdaten/spieler/148153/plus/1?saison=2018",
+            #     "robert-sanchez/leistungsdaten/spieler/403151/plus/1?saison=2018",
+            #     "danny-welbeck/leistungsdaten/spieler/67063/plus/1?saison=2018",
+            #     "jayson-molumby/leistungsdaten/spieler/357656/plus/1?saison=2018",
+            #     "andi-zeqiri/leistungsdaten/spieler/345468/plus/1?saison=2018",
+            #     "ben-white/leistungsdaten/spieler/335721/plus/1?saison=2018",
+            #     "alexis-mac-allister/leistungsdaten/spieler/534033/plus/1?saison=2018",
+            #     "christian-walton/leistungsdaten/spieler/208379/plus/1?saison=2018",
+            #     "solly-march/leistungsdaten/spieler/209212/plus/1?saison=2018",
+            #     "bernardo/leistungsdaten/spieler/364258/plus/1?saison=2018",
+            #     "aaron-connolly/leistungsdaten/spieler/434207/plus/1?saison=2018",
+            #     "joel-veltman/leistungsdaten/spieler/111195/plus/1?saison=2018",
+            #     "leandro-trossard/leistungsdaten/spieler/144028/plus/1?saison=2018",
+            #     "davy-propper/leistungsdaten/spieler/79027/plus/1?saison=2018",
+            #     "jose-izquierdo/leistungsdaten/spieler/147094/plus/1?saison=2018",
+            #     "florin-andone/leistungsdaten/spieler/177065/plus/1?saison=2018",
+            #     "jason-steele/leistungsdaten/spieler/73564/plus/1?saison=2018",
+            #     "tariq-lamptey/leistungsdaten/spieler/504148/plus/1?saison=2018"
+            # ],
+            # "/crystal-palace/startseite/verein/873/saison_id/2020": [
+            #     "andros-townsend/leistungsdaten/spieler/61842/plus/1?saison=2018",
+            #     "stephen-henderson/leistungsdaten/spieler/52191/plus/1?saison=2018",
+            #     "joel-ward/leistungsdaten/spieler/92572/plus/1?saison=2018",
+            #     "martin-kelly/leistungsdaten/spieler/78959/plus/1?saison=2018",
+            #     "nathaniel-clyne/leistungsdaten/spieler/85177/plus/1?saison=2018",
+            #     "james-mccarthy/leistungsdaten/spieler/45333/plus/1?saison=2018",
+            #     "wayne-hennessey/leistungsdaten/spieler/45494/plus/1?saison=2018",
+            #     "jairo-riedewald/leistungsdaten/spieler/241481/plus/1?saison=2018",
+            #     "gary-cahill/leistungsdaten/spieler/27511/plus/1?saison=2018",
+            #     "max-meyer/leistungsdaten/spieler/146164/plus/1?saison=2018",
+            #     "patrick-van-aanholt/leistungsdaten/spieler/52119/plus/1?saison=2018",
+            #     "christian-benteke/leistungsdaten/spieler/50201/plus/1?saison=2018",
+            #     "connor-wickham/leistungsdaten/spieler/95435/plus/1?saison=2018",
+            #     "luka-milivojevic/leistungsdaten/spieler/74300/plus/1?saison=2018",
+            #     "scott-dann/leistungsdaten/spieler/62688/plus/1?saison=2018",
+            #     "sam-woods/leistungsdaten/spieler/611650/plus/1?saison=2018",
+            #     "tyrick-mitchell/leistungsdaten/spieler/730893/plus/1?saison=2018",
+            #     "james-tomkins/leistungsdaten/spieler/61592/plus/1?saison=2018",
+            #     "vicente-guaita/leistungsdaten/spieler/64399/plus/1?saison=2018",
+            #     "michy-batshuayi/leistungsdaten/spieler/179184/plus/1?saison=2018",
+            #     "jordan-ayew/leistungsdaten/spieler/108354/plus/1?saison=2018",
+            #     "eberechi-eze/leistungsdaten/spieler/479999/plus/1?saison=2018",
+            #     "cheikhou-kouyate/leistungsdaten/spieler/66934/plus/1?saison=2018",
+            #     "wilfried-zaha/leistungsdaten/spieler/145988/plus/1?saison=2018",
+            #     "jack-butland/leistungsdaten/spieler/128899/plus/1?saison=2018",
+            #     "james-mcarthur/leistungsdaten/spieler/41416/plus/1?saison=2018",
+            #     "jeffrey-schlupp/leistungsdaten/spieler/157506/plus/1?saison=2018",
+            #     "nathan-ferguson/leistungsdaten/spieler/507254/plus/1?saison=2018",
+            #     "mamadou-sakho/leistungsdaten/spieler/47713/plus/1?saison=2018"
+            # ],
+            # "/fc-arsenal/startseite/verein/11/saison_id/2020": [
+            #     "bernd-leno/leistungsdaten/spieler/72476/plus/1?saison=2018",
+            #     "matt-macey/leistungsdaten/spieler/220491/plus/1?saison=2018",
+            #     "mesut-ozil/leistungsdaten/spieler/35664/plus/1?saison=2018",
+            #     "kieran-tierney/leistungsdaten/spieler/300716/plus/1?saison=2018",
+            #     "david-luiz/leistungsdaten/spieler/46741/plus/1?saison=2018",
+            #     "sokratis/leistungsdaten/spieler/34322/plus/1?saison=2018",
+            #     "thomas-partey/leistungsdaten/spieler/230784/plus/1?saison=2018",
+            #     "hector-bellerin/leistungsdaten/spieler/191217/plus/1?saison=2018",
+            #     "joe-willock/leistungsdaten/spieler/340329/plus/1?saison=2018",
+            #     "cedric-soares/leistungsdaten/spieler/112988/plus/1?saison=2018",
+            #     "nicolas-pepe/leistungsdaten/spieler/343052/plus/1?saison=2018",
+            #     "bukayo-saka/leistungsdaten/spieler/433177/plus/1?saison=2018",
+            #     "gabriel/leistungsdaten/spieler/435338/plus/1?saison=2018",
+            #     "calum-chambers/leistungsdaten/spieler/215118/plus/1?saison=2018",
+            #     "dani-ceballos/leistungsdaten/spieler/319745/plus/1?saison=2018",
+            #     "pierre-emerick-aubameyang/leistungsdaten/spieler/58864/plus/1?saison=2018",
+            #     "rob-holding/leistungsdaten/spieler/253341/plus/1?saison=2018",
+            #     "gabriel-martinelli/leistungsdaten/spieler/655488/plus/1?saison=2018",
+            #     "alexandre-lacazette/leistungsdaten/spieler/93720/plus/1?saison=2018",
+            #     "shkodran-mustafi/leistungsdaten/spieler/88590/plus/1?saison=2018",
+            #     "runar-alex-runarsson/leistungsdaten/spieler/205657/plus/1?saison=2018",
+            #     "willian/leistungsdaten/spieler/52769/plus/1?saison=2018",
+            #     "mohamed-elneny/leistungsdaten/spieler/160438/plus/1?saison=2018",
+            #     "reiss-nelson/leistungsdaten/spieler/340325/plus/1?saison=2018",
+            #     "ainsley-maitland-niles/leistungsdaten/spieler/285845/plus/1?saison=2018",
+            #     "eddie-nketiah/leistungsdaten/spieler/340324/plus/1?saison=2018",
+            #     "pablo-mari/leistungsdaten/spieler/210178/plus/1?saison=2018",
+            #     "omar-rekik/leistungsdaten/spieler/578292/plus/1?saison=2018",
+            #     "granit-xhaka/leistungsdaten/spieler/111455/plus/1?saison=2018",
+            #     "sead-kolasinac/leistungsdaten/spieler/94005/plus/1?saison=2018",
+            #     "emile-smith-rowe/leistungsdaten/spieler/392765/plus/1?saison=2018",
+            #     "william-saliba/leistungsdaten/spieler/495666/plus/1?saison=2018"
+            # ],
+            # "/fc-burnley/startseite/verein/1132/saison_id/2020": [
+            #     "bobby-thomas/leistungsdaten/spieler/788283/plus/1?saison=2018",
+            #     "kevin-long/leistungsdaten/spieler/111114/plus/1?saison=2018",
+            #     "lukas-jensen/leistungsdaten/spieler/581890/plus/1?saison=2018",
+            #     "matthew-lowton/leistungsdaten/spieler/102258/plus/1?saison=2018",
+            #     "josh-brownhill/leistungsdaten/spieler/293569/plus/1?saison=2018",
+            #     "phil-bardsley/leistungsdaten/spieler/15773/plus/1?saison=2018",
+            #     "james-tarkowski/leistungsdaten/spieler/173504/plus/1?saison=2018",
+            #     "jack-cork/leistungsdaten/spieler/40613/plus/1?saison=2018",
+            #     "dwight-mcneil/leistungsdaten/spieler/584769/plus/1?saison=2018",
+            #     "robbie-brady/leistungsdaten/spieler/128229/plus/1?saison=2018",
+            #     "bailey-peacock-farrell/leistungsdaten/spieler/410436/plus/1?saison=2018",
+            #     "johann-berg-gudmundsson/leistungsdaten/spieler/89231/plus/1?saison=2018",
+            #     "jay-rodriguez/leistungsdaten/spieler/53360/plus/1?saison=2018",
+            #     "richard-nartey/leistungsdaten/spieler/396861/plus/1?saison=2018",
+            #     "ashley-westwood/leistungsdaten/spieler/91317/plus/1?saison=2018",
+            #     "ashley-barnes/leistungsdaten/spieler/63200/plus/1?saison=2018",
+            #     "will-norris/leistungsdaten/spieler/285033/plus/1?saison=2018",
+            #     "dale-stephens/leistungsdaten/spieler/49755/plus/1?saison=2018",
+            #     "jimmy-dunne/leistungsdaten/spieler/396174/plus/1?saison=2018",
+            #     "nick-pope/leistungsdaten/spieler/192080/plus/1?saison=2018",
+            #     "erik-pieters/leistungsdaten/spieler/43763/plus/1?saison=2018",
+            #     "matej-vydra/leistungsdaten/spieler/101500/plus/1?saison=2018",
+            #     "chris-wood/leistungsdaten/spieler/108725/plus/1?saison=2018",
+            #     "ben-mee/leistungsdaten/spieler/74810/plus/1?saison=2018",
+            #     "josh-benson/leistungsdaten/spieler/397102/plus/1?saison=2018",
+            #     "mace-goodridge/leistungsdaten/spieler/382552/plus/1?saison=2018",
+            #     "charlie-taylor/leistungsdaten/spieler/195633/plus/1?saison=2018"
+            # ],
+            # "/fc-chelsea/startseite/verein/631/saison_id/2020": [
+            #     "billy-gilmour/leistungsdaten/spieler/423744/plus/1?saison=2018",
+            #     "antonio-rudiger/leistungsdaten/spieler/86202/plus/1?saison=2018",
+            #     "ben-chilwell/leistungsdaten/spieler/316125/plus/1?saison=2018",
+            #     "karlo-ziger/leistungsdaten/spieler/526258/plus/1?saison=2018",
+            #     "cesar-azpilicueta/leistungsdaten/spieler/57500/plus/1?saison=2018",
+            #     "callum-hudson-odoi/leistungsdaten/spieler/392768/plus/1?saison=2018",
+            #     "willy-caballero/leistungsdaten/spieler/19948/plus/1?saison=2018",
+            #     "timo-werner/leistungsdaten/spieler/170527/plus/1?saison=2018",
+            #     "emerson/leistungsdaten/spieler/181778/plus/1?saison=2018",
+            #     "kepa/leistungsdaten/spieler/192279/plus/1?saison=2018",
+            #     "reece-james/leistungsdaten/spieler/472423/plus/1?saison=2018",
+            #     "kai-havertz/leistungsdaten/spieler/309400/plus/1?saison=2018",
+            #     "andreas-christensen/leistungsdaten/spieler/196948/plus/1?saison=2018",
+            #     "christian-pulisic/leistungsdaten/spieler/315779/plus/1?saison=2018",
+            #     "ngolo-kante/leistungsdaten/spieler/225083/plus/1?saison=2018",
+            #     "fikayo-tomori/leistungsdaten/spieler/303254/plus/1?saison=2018",
+            #     "mateo-kovacic/leistungsdaten/spieler/51471/plus/1?saison=2018",
+            #     "abdul-rahman-baba/leistungsdaten/spieler/224884/plus/1?saison=2018",
+            #     "mason-mount/leistungsdaten/spieler/346483/plus/1?saison=2018",
+            #     "marcos-alonso/leistungsdaten/spieler/112515/plus/1?saison=2018",
+            #     "jorginho/leistungsdaten/spieler/102017/plus/1?saison=2018",
+            #     "danny-drinkwater/leistungsdaten/spieler/73491/plus/1?saison=2018",
+            #     "edouard-mendy/leistungsdaten/spieler/442531/plus/1?saison=2018",
+            #     "hakim-ziyech/leistungsdaten/spieler/217111/plus/1?saison=2018",
+            #     "olivier-giroud/leistungsdaten/spieler/82442/plus/1?saison=2018",
+            #     "charly-musonda-jr-/leistungsdaten/spieler/177862/plus/1?saison=2018",
+            #     "kurt-zouma/leistungsdaten/spieler/157509/plus/1?saison=2018",
+            #     "thiago-silva/leistungsdaten/spieler/29241/plus/1?saison=2018",
+            #     "tammy-abraham/leistungsdaten/spieler/331726/plus/1?saison=2018"
+            # ],
+            # "/fc-everton/startseite/verein/29/saison_id/2020": [
+            #     "niels-nkounkou/leistungsdaten/spieler/591193/plus/1?saison=2018",
+            #     "james-rodriguez/leistungsdaten/spieler/88103/plus/1?saison=2018",
+            #     "andre-gomes/leistungsdaten/spieler/221025/plus/1?saison=2018",
+            #     "jordan-pickford/leistungsdaten/spieler/130164/plus/1?saison=2018",
+            #     "fabian-delph/leistungsdaten/spieler/50362/plus/1?saison=2018",
+            #     "mason-holgate/leistungsdaten/spieler/348623/plus/1?saison=2018",
+            #     "alex-iwobi/leistungsdaten/spieler/242631/plus/1?saison=2018",
+            #     "robin-olsen/leistungsdaten/spieler/75458/plus/1?saison=2018",
+            #     "jonjoe-kenny/leistungsdaten/spieler/258883/plus/1?saison=2018",
+            #     "cenk-tosun/leistungsdaten/spieler/45671/plus/1?saison=2018",
+            #     "seamus-coleman/leistungsdaten/spieler/68390/plus/1?saison=2018",
+            #     "gylfi-sigurdsson/leistungsdaten/spieler/90466/plus/1?saison=2018",
+            #     "allan/leistungsdaten/spieler/126422/plus/1?saison=2018",
+            #     "ben-godfrey/leistungsdaten/spieler/343475/plus/1?saison=2018",
+            #     "richarlison/leistungsdaten/spieler/378710/plus/1?saison=2018",
+            #     "yannick-bolasie/leistungsdaten/spieler/75471/plus/1?saison=2018",
+            #     "jarrad-branthwaite/leistungsdaten/spieler/661053/plus/1?saison=2018",
+            #     "abdoulaye-doucoure/leistungsdaten/spieler/127187/plus/1?saison=2018",
+            #     "jonas-lossl/leistungsdaten/spieler/48870/plus/1?saison=2018",
+            #     "beni-baningime/leistungsdaten/spieler/399369/plus/1?saison=2018",
+            #     "anthony-gordon/leistungsdaten/spieler/503733/plus/1?saison=2018",
+            #     "yerry-mina/leistungsdaten/spieler/289446/plus/1?saison=2018",
+            #     "dominic-calvert-lewin/leistungsdaten/spieler/306024/plus/1?saison=2018",
+            #     "jean-philippe-gbamin/leistungsdaten/spieler/182894/plus/1?saison=2018",
+            #     "joao-virginia/leistungsdaten/spieler/329813/plus/1?saison=2018",
+            #     "michael-keane/leistungsdaten/spieler/118534/plus/1?saison=2018",
+            #     "lucas-digne/leistungsdaten/spieler/126664/plus/1?saison=2018",
+            #     "bernard/leistungsdaten/spieler/175169/plus/1?saison=2018",
+            #     "tom-davies/leistungsdaten/spieler/314210/plus/1?saison=2018"
+            # ],
+            # "/fc-fulham/startseite/verein/931/saison_id/2020": [
+            #     "ola-aina/leistungsdaten/spieler/236490/plus/1?saison=2018",
+            #     "mario-lemina/leistungsdaten/spieler/170934/plus/1?saison=2018",
+            #     "michael-hector/leistungsdaten/spieler/157490/plus/1?saison=2018",
+            #     "kevin-mcdonald/leistungsdaten/spieler/36529/plus/1?saison=2018",
+            #     "terence-kongolo/leistungsdaten/spieler/152651/plus/1?saison=2018",
+            #     "ivan-cavaleiro/leistungsdaten/spieler/149716/plus/1?saison=2018",
+            #     "aleksandar-mitrovic/leistungsdaten/spieler/51152/plus/1?saison=2018",
+            #     "fabri/leistungsdaten/spieler/45882/plus/1?saison=2018",
+            #     "joachim-andersen/leistungsdaten/spieler/260827/plus/1?saison=2018",
+            #     "jean-michael-seri/leistungsdaten/spieler/178614/plus/1?saison=2018",
+            #     "bobby-reid/leistungsdaten/spieler/186186/plus/1?saison=2018",
+            #     "neeskens-kebano/leistungsdaten/spieler/111422/plus/1?saison=2018",
+            #     "kenny-tete/leistungsdaten/spieler/206746/plus/1?saison=2018",
+            #     "antonee-robinson/leistungsdaten/spieler/349701/plus/1?saison=2018",
+            #     "tim-ream/leistungsdaten/spieler/145466/plus/1?saison=2018",
+            #     "aboubakar-kamara/leistungsdaten/spieler/290951/plus/1?saison=2018",
+            #     "tom-cairney/leistungsdaten/spieler/123275/plus/1?saison=2018",
+            #     "ruben-loftus-cheek/leistungsdaten/spieler/202886/plus/1?saison=2018",
+            #     "josh-onomah/leistungsdaten/spieler/243589/plus/1?saison=2018",
+            #     "harrison-reed/leistungsdaten/spieler/226973/plus/1?saison=2018",
+            #     "joe-bryan/leistungsdaten/spieler/194460/plus/1?saison=2018",
+            #     "stefan-johansen/leistungsdaten/spieler/62008/plus/1?saison=2018",
+            #     "tosin-adarabioyo/leistungsdaten/spieler/258878/plus/1?saison=2018",
+            #     "ademola-lookman/leistungsdaten/spieler/406040/plus/1?saison=2018",
+            #     "denis-odoi/leistungsdaten/spieler/59641/plus/1?saison=2018",
+            #     "marek-rodak/leistungsdaten/spieler/186901/plus/1?saison=2018",
+            #     "andre-zambo-anguissa/leistungsdaten/spieler/354361/plus/1?saison=2018",
+            #     "alphonse-areola/leistungsdaten/spieler/120629/plus/1?saison=2018",
+            #     "maxime-le-marchand/leistungsdaten/spieler/60558/plus/1?saison=2018"
+            # ],
+            "/fc-liverpool/startseite/verein/31/saison_id/2020": [
+            #     "fabinho/leistungsdaten/spieler/225693/plus/1?saison=2018",
+            #     "takumi-minamino/leistungsdaten/spieler/165793/plus/1?saison=2018",
+            #     "nathaniel-phillips/leistungsdaten/spieler/371814/plus/1?saison=2018",
+            #     "mohamed-salah/leistungsdaten/spieler/148455/plus/1?saison=2018",
+            #     "sadio-mane/leistungsdaten/spieler/200512/plus/1?saison=2018",
+            #     "joel-matip/leistungsdaten/spieler/82105/plus/1?saison=2018",
+            #     "joe-gomez/leistungsdaten/spieler/256178/plus/1?saison=2018",
+            #     "georginio-wijnaldum/leistungsdaten/spieler/49499/plus/1?saison=2018",
+            #     "roberto-firmino/leistungsdaten/spieler/131789/plus/1?saison=2018",
+            #     "jordan-henderson/leistungsdaten/spieler/61651/plus/1?saison=2018",
+            #     "sepp-van-den-berg/leistungsdaten/spieler/541231/plus/1?saison=2018",
+            #     "alisson/leistungsdaten/spieler/105470/plus/1?saison=2018",
+            #     "alex-oxlade-chamberlain/leistungsdaten/spieler/143424/plus/1?saison=2018",
+            #     "konstantinos-tsimikas/leistungsdaten/spieler/338070/plus/1?saison=2018",
+            #     "rhys-williams/leistungsdaten/spieler/503679/plus/1?saison=2018",
+            #     "thiago/leistungsdaten/spieler/60444/plus/1?saison=2018",
+            #     "xherdan-shaqiri/leistungsdaten/spieler/86792/plus/1?saison=2018",
+            #     "adrian/leistungsdaten/spieler/71271/plus/1?saison=2018",
+            #     "neco-williams/leistungsdaten/spieler/503680/plus/1?saison=2018",
+            #     "divock-origi/leistungsdaten/spieler/148368/plus/1?saison=2018",
+                "diogo-jota/leistungsdaten/spieler/340950/plus/1?saison=2018",
+            #     "andrew-robertson/leistungsdaten/spieler/234803/plus/1?saison=2018",
+            #     "virgil-van-dijk/leistungsdaten/spieler/139208/plus/1?saison=2018",
+            #     "curtis-jones/leistungsdaten/spieler/433188/plus/1?saison=2018",
+            #     "trent-alexander-arnold/leistungsdaten/spieler/314353/plus/1?saison=2018",
+            #     "caoimhin-kelleher/leistungsdaten/spieler/340918/plus/1?saison=2018",
+            #     "naby-keita/leistungsdaten/spieler/302215/plus/1?saison=2018",
+            #     "james-milner/leistungsdaten/spieler/3333/plus/1?saison=2018"
+            # ],
+            # "/fc-southampton/startseite/verein/180/saison_id/2020": [
+            #     "sam-mcqueen/leistungsdaten/spieler/226968/plus/1?saison=2018",
+            #     "oriol-romeu/leistungsdaten/spieler/66100/plus/1?saison=2018",
+            #     "moussa-djenepo/leistungsdaten/spieler/485424/plus/1?saison=2018",
+            #     "ryan-bertrand/leistungsdaten/spieler/40611/plus/1?saison=2018",
+            #     "harry-lewis/leistungsdaten/spieler/341043/plus/1?saison=2018",
+            #     "shane-long/leistungsdaten/spieler/37304/plus/1?saison=2018",
+            #     "che-adams/leistungsdaten/spieler/346779/plus/1?saison=2018",
+            #     "michael-obafemi/leistungsdaten/spieler/444208/plus/1?saison=2018",
+            #     "will-smallbone/leistungsdaten/spieler/444211/plus/1?saison=2018",
+            #     "nathan-redmond/leistungsdaten/spieler/129078/plus/1?saison=2018",
+            #     "james-ward-prowse/leistungsdaten/spieler/181579/plus/1?saison=2018",
+            #     "jannik-vestergaard/leistungsdaten/spieler/99331/plus/1?saison=2018",
+            #     "kyle-walker-peters/leistungsdaten/spieler/341051/plus/1?saison=2018",
+            #     "brandon-williams/leistungsdaten/spieler/507700/plus/1?saison=2018",
+            #     "danny-ings/leistungsdaten/spieler/134294/plus/1?saison=2018",
+            #     "dan-nlundulu/leistungsdaten/spieler/346482/plus/1?saison=2018",
+            #     "jan-bednarek/leistungsdaten/spieler/243028/plus/1?saison=2018",
+            #     "ibrahima-diallo/leistungsdaten/spieler/413032/plus/1?saison=2018",
+            #     "fraser-forster/leistungsdaten/spieler/52570/plus/1?saison=2018",
+            #     "jake-vokins/leistungsdaten/spieler/421424/plus/1?saison=2018",
+            #     "stuart-armstrong/leistungsdaten/spieler/130765/plus/1?saison=2018",
+            #     "mohammed-salisu/leistungsdaten/spieler/563963/plus/1?saison=2018",
+            #     "jack-stephens/leistungsdaten/spieler/163744/plus/1?saison=2018",
+            #     "yan-valery/leistungsdaten/spieler/406008/plus/1?saison=2018",
+            #     "theo-walcott/leistungsdaten/spieler/33713/plus/1?saison=2018",
+            #     "nathan-tella/leistungsdaten/spieler/340322/plus/1?saison=2018",
+            #     "alex-mccarthy/leistungsdaten/spieler/95976/plus/1?saison=2018"
+            ],
+            # "/leeds-united/startseite/verein/399/saison_id/2020": [
+            #     "robin-koch/leistungsdaten/spieler/328784/plus/1?saison=2018",
+            #     "ezgjan-alioski/leistungsdaten/spieler/129604/plus/1?saison=2018",
+            #     "luke-ayling/leistungsdaten/spieler/67420/plus/1?saison=2018",
+            #     "diego-llorente/leistungsdaten/spieler/246291/plus/1?saison=2018",
+            #     "liam-cooper/leistungsdaten/spieler/75067/plus/1?saison=2018",
+            #     "pablo-hernandez/leistungsdaten/spieler/46220/plus/1?saison=2018",
+            #     "patrick-bamford/leistungsdaten/spieler/183334/plus/1?saison=2018",
+            #     "gaetano-berardi/leistungsdaten/spieler/43084/plus/1?saison=2018",
+            #     "oliver-casey/leistungsdaten/spieler/559800/plus/1?saison=2018",
+            #     "adam-forshaw/leistungsdaten/spieler/121257/plus/1?saison=2018",
+            #     "leif-davis/leistungsdaten/spieler/596446/plus/1?saison=2018",
+            #     "jamie-shackleton/leistungsdaten/spieler/534700/plus/1?saison=2018",
+            #     "pascal-struijk/leistungsdaten/spieler/410708/plus/1?saison=2018",
+            #     "illan-meslier/leistungsdaten/spieler/542586/plus/1?saison=2018",
+            #     "conor-shaughnessy/leistungsdaten/spieler/255626/plus/1?saison=2018",
+            #     "helder-costa/leistungsdaten/spieler/173900/plus/1?saison=2018",
+            #     "stuart-dallas/leistungsdaten/spieler/158764/plus/1?saison=2018",
+            #     "rodrigo/leistungsdaten/spieler/131505/plus/1?saison=2018",
+            #     "jay-roy-grot/leistungsdaten/spieler/360517/plus/1?saison=2018",
+            #     "kiko-casilla/leistungsdaten/spieler/27486/plus/1?saison=2018",
+            #     "raphinha/leistungsdaten/spieler/411295/plus/1?saison=2018",
+            #     "ian-poveda/leistungsdaten/spieler/392763/plus/1?saison=2018",
+            #     "joe-gelhardt/leistungsdaten/spieler/503980/plus/1?saison=2018",
+            #     "mateusz-klich/leistungsdaten/spieler/92738/plus/1?saison=2018",
+            #     "riqui-puig/leistungsdaten/spieler/331511/plus/1?saison=2018",
+            #     "kalvin-phillips/leistungsdaten/spieler/351749/plus/1?saison=2018",
+            #     "ouasim-bouy/leistungsdaten/spieler/126412/plus/1?saison=2018",
+            #     "tyler-roberts/leistungsdaten/spieler/296986/plus/1?saison=2018",
+            #     "jack-harrison/leistungsdaten/spieler/417346/plus/1?saison=2018",
+            #     "elia-caprile/leistungsdaten/spieler/421873/plus/1?saison=2018"
+            # ],
+            # "/leicester-city/startseite/verein/1003/saison_id/2020": [
+            #     "dennis-praet/leistungsdaten/spieler/129588/plus/1?saison=2018",
+            #     "timothy-castagne/leistungsdaten/spieler/262226/plus/1?saison=2018",
+            #     "wilfred-ndidi/leistungsdaten/spieler/274839/plus/1?saison=2018",
+            #     "jamie-vardy/leistungsdaten/spieler/197838/plus/1?saison=2018",
+            #     "hamza-choudhury/leistungsdaten/spieler/341501/plus/1?saison=2018",
+            #     "luke-thomas/leistungsdaten/spieler/505194/plus/1?saison=2018",
+            #     "james-maddison/leistungsdaten/spieler/294057/plus/1?saison=2018",
+            #     "wesley-fofana/leistungsdaten/spieler/475411/plus/1?saison=2018",
+            #     "harvey-barnes/leistungsdaten/spieler/398065/plus/1?saison=2018",
+            #     "eldin-jakupovic/leistungsdaten/spieler/2857/plus/1?saison=2018",
+            #     "marc-albrighton/leistungsdaten/spieler/61560/plus/1?saison=2018",
+            #     "jonny-evans/leistungsdaten/spieler/42412/plus/1?saison=2018",
+            #     "caglar-soyuncu/leistungsdaten/spieler/320141/plus/1?saison=2018",
+            #     "gonzalo-plata/leistungsdaten/spieler/592735/plus/1?saison=2018",
+            #     "ayoze-perez/leistungsdaten/spieler/246968/plus/1?saison=2018",
+            #     "ricardo-pereira/leistungsdaten/spieler/215876/plus/1?saison=2018",
+            #     "daniel-amartey/leistungsdaten/spieler/214056/plus/1?saison=2018",
+            #     "christian-fuchs/leistungsdaten/spieler/6636/plus/1?saison=2018",
+            #     "danny-ward/leistungsdaten/spieler/203026/plus/1?saison=2018",
+            #     "wes-morgan/leistungsdaten/spieler/10003/plus/1?saison=2018",
+            #     "demarai-gray/leistungsdaten/spieler/292417/plus/1?saison=2018",
+            #     "cengiz-under/leistungsdaten/spieler/341647/plus/1?saison=2018",
+            #     "james-justin/leistungsdaten/spieler/413220/plus/1?saison=2018",
+            #     "kelechi-iheanacho/leistungsdaten/spieler/295330/plus/1?saison=2018",
+            #     "kasper-schmeichel/leistungsdaten/spieler/16911/plus/1?saison=2018",
+            #     "youri-tielemans/leistungsdaten/spieler/249565/plus/1?saison=2018",
+            #     "nampalys-mendy/leistungsdaten/spieler/111051/plus/1?saison=2018",
+            #     "islam-slimani/leistungsdaten/spieler/174915/plus/1?saison=2018"
+            # ],
+            # "/manchester-city/startseite/verein/281/saison_id/2020": [
+            #     "raheem-sterling/leistungsdaten/spieler/134425/plus/1?saison=2018",
+            #     "rodri/leistungsdaten/spieler/357565/plus/1?saison=2018",
+            #     "joao-cancelo/leistungsdaten/spieler/182712/plus/1?saison=2018",
+            #     "cole-palmer/leistungsdaten/spieler/568177/plus/1?saison=2018",
+            #     "benjamin-mendy/leistungsdaten/spieler/157495/plus/1?saison=2018",
+            #     "eric-garcia/leistungsdaten/spieler/466794/plus/1?saison=2018",
+            #     "fernandinho/leistungsdaten/spieler/26267/plus/1?saison=2018",
+            #     "riyad-mahrez/leistungsdaten/spieler/171424/plus/1?saison=2018",
+            #     "kevin-de-bruyne/leistungsdaten/spieler/88755/plus/1?saison=2018",
+            #     "john-stones/leistungsdaten/spieler/186590/plus/1?saison=2018",
+            #     "oleksandr-zinchenko/leistungsdaten/spieler/203853/plus/1?saison=2018",
+            #     "gabriel-jesus/leistungsdaten/spieler/363205/plus/1?saison=2018",
+            #     "ruben-dias/leistungsdaten/spieler/258004/plus/1?saison=2018",
+            #     "nathan-ake/leistungsdaten/spieler/177476/plus/1?saison=2018",
+            #     "ilkay-gundogan/leistungsdaten/spieler/53622/plus/1?saison=2018",
+            #     "ferran-torres/leistungsdaten/spieler/398184/plus/1?saison=2018",
+            #     "scott-carson/leistungsdaten/spieler/14555/plus/1?saison=2018",
+            #     "zack-steffen/leistungsdaten/spieler/221624/plus/1?saison=2018",
+            #     "philippe-sandler/leistungsdaten/spieler/340460/plus/1?saison=2018",
+            #     "kyle-walker/leistungsdaten/spieler/95424/plus/1?saison=2018",
+            #     "bernardo-silva/leistungsdaten/spieler/241641/plus/1?saison=2018",
+            #     "ederson/leistungsdaten/spieler/238223/plus/1?saison=2018",
+            #     "sergio-aguero/leistungsdaten/spieler/26399/plus/1?saison=2018",
+            #     "phil-foden/leistungsdaten/spieler/406635/plus/1?saison=2018",
+            #     "aymeric-laporte/leistungsdaten/spieler/176553/plus/1?saison=2018"
+            # ],
+            # "/manchester-united/startseite/verein/985/saison_id/2020": [
+            #     "mason-greenwood/leistungsdaten/spieler/532826/plus/1?saison=2018",
+            #     "luke-shaw/leistungsdaten/spieler/183288/plus/1?saison=2018",
+            #     "sergio-romero/leistungsdaten/spieler/30690/plus/1?saison=2018",
+            #     "lee-grant/leistungsdaten/spieler/13460/plus/1?saison=2018",
+            #     "timothy-fosu-mensah/leistungsdaten/spieler/315131/plus/1?saison=2018",
+            #     "odion-ighalo/leistungsdaten/spieler/62121/plus/1?saison=2018",
+            #     "alex-telles/leistungsdaten/spieler/255755/plus/1?saison=2018",
+            #     "anthony-martial/leistungsdaten/spieler/182877/plus/1?saison=2018",
+            #     "juan-mata/leistungsdaten/spieler/44068/plus/1?saison=2018",
+            #     "paul-pogba/leistungsdaten/spieler/122153/plus/1?saison=2018",
+            #     "edinson-cavani/leistungsdaten/spieler/48280/plus/1?saison=2018",
+            #     "brandon-williams/leistungsdaten/spieler/507700/plus/1?saison=2018",
+            #     "nemanja-matic/leistungsdaten/spieler/74683/plus/1?saison=2018",
+            #     "eric-bailly/leistungsdaten/spieler/286384/plus/1?saison=2018",
+            #     "harry-maguire/leistungsdaten/spieler/177907/plus/1?saison=2018",
+            #     "scott-mctominay/leistungsdaten/spieler/315969/plus/1?saison=2018",
+            #     "david-de-gea/leistungsdaten/spieler/59377/plus/1?saison=2018",
+            #     "victor-lindelof/leistungsdaten/spieler/184573/plus/1?saison=2018",
+            #     "bruno-fernandes/leistungsdaten/spieler/240306/plus/1?saison=2018",
+            #     "fred/leistungsdaten/spieler/191614/plus/1?saison=2018",
+            #     "dean-henderson/leistungsdaten/spieler/258919/plus/1?saison=2018",
+            #     "jesse-lingard/leistungsdaten/spieler/141660/plus/1?saison=2018",
+            #     "axel-tuanzebe/leistungsdaten/spieler/342046/plus/1?saison=2018",
+            #     "daniel-james/leistungsdaten/spieler/319301/plus/1?saison=2018",
+            #     "phil-jones/leistungsdaten/spieler/117996/plus/1?saison=2018",
+            #     "marcos-rojo/leistungsdaten/spieler/93176/plus/1?saison=2018",
+            #     "donny-van-de-beek/leistungsdaten/spieler/288255/plus/1?saison=2018",
+            #     "aaron-wan-bissaka/leistungsdaten/spieler/477758/plus/1?saison=2018",
+            #     "facundo-pellistri/leistungsdaten/spieler/676318/plus/1?saison=2018",
+            #     "marcus-rashford/leistungsdaten/spieler/258923/plus/1?saison=2018"
+            # ],
+            # "/newcastle-united/startseite/verein/762/saison_id/2020": [
+            #     "sean-longstaff/leistungsdaten/spieler/346707/plus/1?saison=2018",
+            #     "jeff-hendrick/leistungsdaten/spieler/148262/plus/1?saison=2018",
+            #     "henri-saivet/leistungsdaten/spieler/51540/plus/1?saison=2018",
+            #     "dwight-gayle/leistungsdaten/spieler/196522/plus/1?saison=2018",
+            #     "miguel-almiron/leistungsdaten/spieler/272999/plus/1?saison=2018",
+            #     "callum-wilson/leistungsdaten/spieler/123682/plus/1?saison=2018",
+            #     "ciaran-clark/leistungsdaten/spieler/98240/plus/1?saison=2018",
+            #     "fabian-schar/leistungsdaten/spieler/135343/plus/1?saison=2018",
+            #     "christian-atsu/leistungsdaten/spieler/186997/plus/1?saison=2018",
+            #     "elliot-anderson/leistungsdaten/spieler/567576/plus/1?saison=2018",
+            #     "mark-gillespie/leistungsdaten/spieler/142389/plus/1?saison=2018",
+            #     "jonjo-shelvey/leistungsdaten/spieler/71292/plus/1?saison=2018",
+            #     "federico-fernandez/leistungsdaten/spieler/85475/plus/1?saison=2018",
+            #     "andy-carroll/leistungsdaten/spieler/48066/plus/1?saison=2018",
+            #     "jamal-lewis/leistungsdaten/spieler/346018/plus/1?saison=2018",
+            #     "ryan-fraser/leistungsdaten/spieler/146795/plus/1?saison=2018",
+            #     "matthew-longstaff/leistungsdaten/spieler/484387/plus/1?saison=2018",
+            #     "allan-saint-maximin/leistungsdaten/spieler/272642/plus/1?saison=2018",
+            #     "martin-dubravka/leistungsdaten/spieler/74960/plus/1?saison=2018",
+            #     "javier-manquillo/leistungsdaten/spieler/162029/plus/1?saison=2018",
+            #     "achraf-lazaar/leistungsdaten/spieler/139173/plus/1?saison=2018",
+            #     "isaac-hayden/leistungsdaten/spieler/206225/plus/1?saison=2018",
+            #     "paul-dummett/leistungsdaten/spieler/170321/plus/1?saison=2018",
+            #     "jacob-murphy/leistungsdaten/spieler/199527/plus/1?saison=2018",
+            #     "marcos-rojo/leistungsdaten/spieler/93176/plus/1?saison=2018",
+            #     "matt-ritchie/leistungsdaten/spieler/92469/plus/1?saison=2018",
+            #     "karl-darlow/leistungsdaten/spieler/99397/plus/1?saison=2018",
+            #     "joelinton/leistungsdaten/spieler/333241/plus/1?saison=2018",
+            #     "emil-krafth/leistungsdaten/spieler/184528/plus/1?saison=2018",
+            #     "deandre-yedlin/leistungsdaten/spieler/255916/plus/1?saison=2018",
+            #     "jamaal-lascelles/leistungsdaten/spieler/183318/plus/1?saison=2018"
+            # ],
+            # "/sheffield-united/startseite/verein/350/saison_id/2020": [
+            #     "lewis-morgan/leistungsdaten/spieler/311351/plus/1?saison=2018",
+            #     "ethan-ampadu/leistungsdaten/spieler/392771/plus/1?saison=2018",
+            #     "ben-osborn/leistungsdaten/spieler/218087/plus/1?saison=2018",
+            #     "phil-jagielka/leistungsdaten/spieler/13520/plus/1?saison=2018",
+            #     "rhian-brewster/leistungsdaten/spieler/406560/plus/1?saison=2018",
+            #     "aaron-ramsdale/leistungsdaten/spieler/427568/plus/1?saison=2018",
+            #     "michael-verrips/leistungsdaten/spieler/288259/plus/1?saison=2018",
+            #     "wes-foderingham/leistungsdaten/spieler/61697/plus/1?saison=2018",
+            #     "jack-robinson/leistungsdaten/spieler/128909/plus/1?saison=2018",
+            #     "john-lundstram/leistungsdaten/spieler/156941/plus/1?saison=2018",
+            #     "jayden-bogle/leistungsdaten/spieler/465717/plus/1?saison=2018",
+            #     "oliver-norwood/leistungsdaten/spieler/73547/plus/1?saison=2018",
+            #     "oliver-burke/leistungsdaten/spieler/341317/plus/1?saison=2018",
+            #     "enda-stevens/leistungsdaten/spieler/85706/plus/1?saison=2018",
+            #     "simon-moore/leistungsdaten/spieler/123536/plus/1?saison=2018",
+            #     "jack-oconnell/leistungsdaten/spieler/212829/plus/1?saison=2018",
+            #     "lys-mousset/leistungsdaten/spieler/291422/plus/1?saison=2018",
+            #     "george-baldock/leistungsdaten/spieler/146690/plus/1?saison=2018",
+            #     "chris-basham/leistungsdaten/spieler/52495/plus/1?saison=2018",
+            #     "oliver-mcburnie/leistungsdaten/spieler/298477/plus/1?saison=2018",
+            #     "max-lowe/leistungsdaten/spieler/258885/plus/1?saison=2018",
+            #     "kean-bryan/leistungsdaten/spieler/214928/plus/1?saison=2018",
+            #     "sander-berge/leistungsdaten/spieler/333014/plus/1?saison=2018",
+            #     "jack-rodwell/leistungsdaten/spieler/57079/plus/1?saison=2018",
+            #     "john-egan/leistungsdaten/spieler/121324/plus/1?saison=2018",
+            #     "david-mcgoldrick/leistungsdaten/spieler/35416/plus/1?saison=2018",
+            #     "billy-sharp/leistungsdaten/spieler/49542/plus/1?saison=2018",
+            #     "john-fleck/leistungsdaten/spieler/54383/plus/1?saison=2018"
+            # ],
+            # "/tottenham-hotspur/startseite/verein/148/saison_id/2020": [
+            #     "danny-rose/leistungsdaten/spieler/50174/plus/1?saison=2018",
+            #     "moussa-sissoko/leistungsdaten/spieler/46001/plus/1?saison=2018",
+            #     "davinson-sanchez/leistungsdaten/spieler/341429/plus/1?saison=2018",
+            #     "lucas-moura/leistungsdaten/spieler/77100/plus/1?saison=2018",
+            #     "hugo-lloris/leistungsdaten/spieler/17965/plus/1?saison=2018",
+            #     "harry-winks/leistungsdaten/spieler/249126/plus/1?saison=2018",
+            #     "eric-dier/leistungsdaten/spieler/175722/plus/1?saison=2018",
+            #     "gedson-fernandes/leistungsdaten/spieler/337800/plus/1?saison=2018",
+            #     "tanguy-ndombele/leistungsdaten/spieler/450936/plus/1?saison=2018",
+            #     "ben-davies/leistungsdaten/spieler/192765/plus/1?saison=2018",
+            #     "sergio-reguilon/leistungsdaten/spieler/282429/plus/1?saison=2018",
+            #     "toby-alderweireld/leistungsdaten/spieler/42710/plus/1?saison=2018",
+            #     "pierre-emile-hojbjerg/leistungsdaten/spieler/167799/plus/1?saison=2018",
+            #     "harry-kane/leistungsdaten/spieler/132098/plus/1?saison=2018",
+            #     "heung-min-son/leistungsdaten/spieler/91845/plus/1?saison=2018",
+            #     "alfie-whiteman/leistungsdaten/spieler/282823/plus/1?saison=2018",
+            #     "joe-rodon/leistungsdaten/spieler/297212/plus/1?saison=2018",
+            #     "paulo-gazzaniga/leistungsdaten/spieler/195488/plus/1?saison=2018",
+            #     "jack-clarke/leistungsdaten/spieler/559794/plus/1?saison=2018",
+            #     "serge-aurier/leistungsdaten/spieler/127032/plus/1?saison=2018",
+            #     "japhet-tanganga/leistungsdaten/spieler/346478/plus/1?saison=2018",
+            #     "joe-hart/leistungsdaten/spieler/40204/plus/1?saison=2018",
+            #     "dele-alli/leistungsdaten/spieler/207929/plus/1?saison=2018",
+            #     "erik-lamela/leistungsdaten/spieler/111630/plus/1?saison=2018",
+            #     "steven-bergwijn/leistungsdaten/spieler/284165/plus/1?saison=2018",
+            #     "gareth-bale/leistungsdaten/spieler/39381/plus/1?saison=2018",
+            #     "matt-doherty/leistungsdaten/spieler/171679/plus/1?saison=2018",
+            #     "carlos-vinicius/leistungsdaten/spieler/525287/plus/1?saison=2018",
+            #     "giovani-lo-celso/leistungsdaten/spieler/348795/plus/1?saison=2018"
+            # ],
+            # "/west-bromwich-albion/startseite/verein/984/saison_id/2020": [
+            #     "charlie-austin/leistungsdaten/spieler/129627/plus/1?saison=2018",
+            #     "kamil-grosicki/leistungsdaten/spieler/37920/plus/1?saison=2018",
+            #     "dara-oshea/leistungsdaten/spieler/401320/plus/1?saison=2018",
+            #     "kyle-edwards/leistungsdaten/spieler/314366/plus/1?saison=2018",
+            #     "sam-johnstone/leistungsdaten/spieler/110864/plus/1?saison=2018",
+            #     "grady-diangana/leistungsdaten/spieler/342048/plus/1?saison=2018",
+            #     "rekeem-harper/leistungsdaten/spieler/456877/plus/1?saison=2018",
+            #     "sam-field/leistungsdaten/spieler/387331/plus/1?saison=2018",
+            #     "jake-livermore/leistungsdaten/spieler/61832/plus/1?saison=2018",
+            #     "conor-townsend/leistungsdaten/spieler/146509/plus/1?saison=2018",
+            #     "matt-phillips/leistungsdaten/spieler/77274/plus/1?saison=2018",
+            #     "kieran-gibbs/leistungsdaten/spieler/44792/plus/1?saison=2018",
+            #     "conor-gallagher/leistungsdaten/spieler/488362/plus/1?saison=2018",
+            #     "kyle-bartley/leistungsdaten/spieler/67421/plus/1?saison=2018",
+            #     "branislav-ivanovic/leistungsdaten/spieler/36827/plus/1?saison=2018",
+            #     "matheus-pereira/leistungsdaten/spieler/225984/plus/1?saison=2018",
+            #     "jonathan-bond/leistungsdaten/spieler/130768/plus/1?saison=2018",
+            #     "darnell-furlong/leistungsdaten/spieler/351755/plus/1?saison=2018",
+            #     "david-button/leistungsdaten/spieler/61814/plus/1?saison=2018",
+            #     "hal-robson-kanu/leistungsdaten/spieler/65976/plus/1?saison=2018",
+            #     "romaine-sawyers/leistungsdaten/spieler/113036/plus/1?saison=2018",
+            #     "semi-ajayi/leistungsdaten/spieler/251912/plus/1?saison=2018",
+            #     "phil-jones/leistungsdaten/spieler/117996/plus/1?saison=2018",
+            #     "karlan-grant/leistungsdaten/spieler/314183/plus/1?saison=2018",
+            #     "filip-krovinovic/leistungsdaten/spieler/261988/plus/1?saison=2018",
+            #     "callum-robinson/leistungsdaten/spieler/183299/plus/1?saison=2018",
+            #     "lee-peltier/leistungsdaten/spieler/38094/plus/1?saison=2018",
+            #     "cedric-kipre/leistungsdaten/spieler/364827/plus/1?saison=2018"
+            # ],
+            # "/west-ham-united/startseite/verein/379/saison_id/2020": [
+            #     "jarrod-bowen/leistungsdaten/spieler/314875/plus/1?saison=2018",
+            #     "vladimir-coufal/leistungsdaten/spieler/157672/plus/1?saison=2018",
+            #     "ryan-fredericks/leistungsdaten/spieler/140191/plus/1?saison=2018",
+            #     "michail-antonio/leistungsdaten/spieler/104124/plus/1?saison=2018",
+            #     "arthur-masuaku/leistungsdaten/spieler/181380/plus/1?saison=2018",
+            #     "tomas-soucek/leistungsdaten/spieler/283628/plus/1?saison=2018",
+            #     "declan-rice/leistungsdaten/spieler/357662/plus/1?saison=2018",
+            #     "david-martin/leistungsdaten/spieler/15930/plus/1?saison=2018",
+            #     "lukasz-fabianski/leistungsdaten/spieler/29692/plus/1?saison=2018",
+            #     "mark-noble/leistungsdaten/spieler/31835/plus/1?saison=2018",
+            #     "oleksandr-zinchenko/leistungsdaten/spieler/203853/plus/1?saison=2018",
+            #     "angelo-ogbonna/leistungsdaten/spieler/48002/plus/1?saison=2018",
+            #     "manuel-lanzini/leistungsdaten/spieler/135853/plus/1?saison=2018",
+            #     "darren-randolph/leistungsdaten/spieler/51321/plus/1?saison=2018",
+            #     "said-benrahma/leistungsdaten/spieler/290532/plus/1?saison=2018",
+            #     "andriy-yarmolenko/leistungsdaten/spieler/69015/plus/1?saison=2018",
+            #     "robert-snodgrass/leistungsdaten/spieler/22614/plus/1?saison=2018",
+            #     "craig-dawson/leistungsdaten/spieler/121477/plus/1?saison=2018",
+            #     "fabian-balbuena/leistungsdaten/spieler/285548/plus/1?saison=2018",
+            #     "issa-diop/leistungsdaten/spieler/272622/plus/1?saison=2018",
+            #     "sebastien-haller/leistungsdaten/spieler/181375/plus/1?saison=2018",
+            #     "pablo-fornals/leistungsdaten/spieler/357885/plus/1?saison=2018",
+            #     "ben-johnson/leistungsdaten/spieler/468002/plus/1?saison=2018",
+            #     "aaron-cresswell/leistungsdaten/spieler/92571/plus/1?saison=2018"
+            # ],
+            # "/wolverhampton-wanderers/startseite/verein/543/saison_id/2020": [
+            #     "romain-saiss/leistungsdaten/spieler/204198/plus/1?saison=2018",
+            #     "fabio-silva/leistungsdaten/spieler/505653/plus/1?saison=2018",
+            #     "conor-coady/leistungsdaten/spieler/128901/plus/1?saison=2018",
+            #     "rui-patricio/leistungsdaten/spieler/45026/plus/1?saison=2018",
+            #     "john-ruddy/leistungsdaten/spieler/29712/plus/1?saison=2018",
+            #     "marcal/leistungsdaten/spieler/137745/plus/1?saison=2018",
+            #     "max-kilman/leistungsdaten/spieler/525247/plus/1?saison=2018",
+            #     "theo-corbeanu/leistungsdaten/spieler/568535/plus/1?saison=2018",
+            #     "luke-matheson/leistungsdaten/spieler/614084/plus/1?saison=2018",
+            #     "daniel-podence/leistungsdaten/spieler/256977/plus/1?saison=2018",
+            #     "oskar-buur/leistungsdaten/spieler/293283/plus/1?saison=2018",
+            #     "adama-traore/leistungsdaten/spieler/204103/plus/1?saison=2018",
+            #     "taylor-perry/leistungsdaten/spieler/502073/plus/1?saison=2018",
+            #     "leander-dendoncker/leistungsdaten/spieler/168157/plus/1?saison=2018",
+            #     "nelson-semedo/leistungsdaten/spieler/231572/plus/1?saison=2018",
+            #     "roderick-miranda/leistungsdaten/spieler/87856/plus/1?saison=2018",
+            #     "owen-otasowie/leistungsdaten/spieler/511815/plus/1?saison=2018",
+            #     "jonny-otto/leistungsdaten/spieler/175446/plus/1?saison=2018",
+            #     "pedro-neto/leistungsdaten/spieler/487465/plus/1?saison=2018",
+            #     "divock-origi/leistungsdaten/spieler/148368/plus/1?saison=2018",
+            #     "willy-boly/leistungsdaten/spieler/142310/plus/1?saison=2018",
+            #     "joao-moutinho/leistungsdaten/spieler/29364/plus/1?saison=2018",
+            #     "renat-dadashov/leistungsdaten/spieler/335461/plus/1?saison=2018",
+            #     "rayan-ait-nouri/leistungsdaten/spieler/578391/plus/1?saison=2018",
+            #     "raul-jimenez/leistungsdaten/spieler/206040/plus/1?saison=2018",
+            #     "ki-jana-hoever/leistungsdaten/spieler/485583/plus/1?saison=2018",
+            #     "vitinha/leistungsdaten/spieler/487469/plus/1?saison=2018",
+            #     "ruben-neves/leistungsdaten/spieler/225161/plus/1?saison=2018"
+            # ]
+        },
+        # "/1-hnl/startseite/wettbewerb/KR1": {},
+        # "/allsvenskan/startseite/wettbewerb/SE1": {},
+        # "/bundesliga/startseite/wettbewerb/A1": {},
+        # "/bundesliga/startseite/wettbewerb/L1": {},
+        # "/eliteserien/startseite/wettbewerb/NO1": {},
+        # "/eredivisie/startseite/wettbewerb/NL1": {},
+        # "/fortuna-liga/startseite/wettbewerb/TS1": {},
+        # "/jupiler-pro-league/startseite/wettbewerb/BE1": {},
+        # "/laliga/startseite/wettbewerb/ES1": {},
+        # "/liga-1/startseite/wettbewerb/RO1": {},
+        # "/liga-nos/startseite/wettbewerb/PO1": {},
+        # "/ligat-haal/startseite/wettbewerb/ISR1": {},
+        # "/ligue-1/startseite/wettbewerb/FR1": {},
+        # "/pko-ekstraklasa/startseite/wettbewerb/PL1": {},
+        # "/premier-liga/startseite/wettbewerb/RU1": {},
+        # "/premier-liga/startseite/wettbewerb/UKR1": {},
+        # "/protathlima-cyta/startseite/wettbewerb/ZYP1": {},
+        # "/scottish-premiership/startseite/wettbewerb/SC1": {},
+        # "/serie-a/startseite/wettbewerb/IT1": {},
+        # "/super-league-1/startseite/wettbewerb/GR1": {},
+        # "/super-league/startseite/wettbewerb/C1": {},
+        # "/super-lig/startseite/wettbewerb/TR1": {},
+        # "/super-liga-srbije/startseite/wettbewerb/SER1": {},
+        # "/superligaen/startseite/wettbewerb/DK1": {}
+    }
+}
